# Parallelize groupby calculations