Testing on my laptop (8 cores) with the following script:
```python
import datatable as dt
import random
import time
from datatable import sum, f
n = 10**8
col1 = [random.randint(1, 100) for _ in range(1000)] * (n // 1000)
col2 = [random.random() for _ in range(20000)] * (n // 20000)
assert len(col1) == len(col2) == n

fr = dt.Frame(A=col1, B=col2)
for _ in range(5):
    t0 = time.time()
    fg = fr[:, sum(f.B), f.A]
    t1 = time.time()
    print("Time taken: %.3fs" % (t1 - t0))
```

Before this change, timings were:
```
Time taken: 2.600s
Time taken: 2.184s
Time taken: 2.140s
Time taken: 2.196s
Time taken: 2.196s
```

After this change, timings are:
```
Time taken: 1.383s
Time taken: 1.028s
Time taken: 1.066s
Time taken: 1.040s
Time taken: 1.051s
```

Closes #1284